% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{all.abundance}
\alias{all.abundance}
\title{K-mer abundances for density calculations}
\format{
An object of class \code{list} of length 3.
}
\usage{
all.abundance
}
\description{
Nested list of k-mer counts keyed by
\code{BSgenome.Hsapiens.1000genomes.hs37d5} / \code{BSgenome.Hsapiens.UCSC.hg38} /
\code{BSgenome.Mmusculus.UCSC.mm10}, then by \code{exome} / \code{transcript} / \code{genome},
then by catalog size as a string (\code{"78"}, \code{"96"}, \code{"136"}, \code{"144"}, \code{"192"},
\code{"1536"}). Each leaf value is a named integer vector: k-mer → count.
}
\details{
Used by \code{\link[=transform_catalog]{transform_catalog()}} to convert counts to density (mutations per
megabase of context).
}
\examples{
all.abundance$BSgenome.Hsapiens.UCSC.hg38$transcript$`144`[1:5]

}
\keyword{datasets}
