% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_attr.R
\name{catalog_attrs}
\alias{catalog_attrs}
\title{Report the attributes of an mSigSpectra catalog}
\usage{
catalog_attrs(x)
}
\arguments{
\item{x}{A catalog.}
}
\value{
A named list with elements \code{type}, \code{counts_or_density},
\code{ref_genome}, \code{region}, \code{abundance}.
}
\description{
Report the attributes of an mSigSpectra catalog
}
