% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_476_type_ids_to_open_intervals.R
\name{change_476_type_ids_to_open_intervals}
\alias{change_476_type_ids_to_open_intervals}
\title{Change 476-type indel category identifiers to use right-open repeat intervals}
\usage{
change_476_type_ids_to_open_intervals(type_476_indel_type_identifiers)
}
\arguments{
\item{type_476_indel_type_identifiers}{Character vector of 476-type indel
category identifiers, e.g. as returned by
\code{categorize_indels_in_vcf()}.}
}
\value{
Character vector the same length as the input, with \verb{:R(X,9)} at the
end of each string replaced by \verb{:R(X,)}.
}
\description{
Replaces bounded repeat-count suffixes like \verb{:R(X,9)} with right-open
equivalents like \verb{:R(X,)}, making the classification agnostic as to whether
repeats longer than 9 were discarded from the input data.
}
\examples{
change_476_type_ids_to_open_intervals(
  c("Del(C):Ins(C):R(5,9)", "Del(T):R(3,5)", "Ins(C):R(5,9)")
)

}
