% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_names.R
\name{check_and_fix_chrom_names}
\alias{check_and_fix_chrom_names}
\title{Check and, if possible, correct the chromosome names in a VCF data.frame}
\usage{
check_and_fix_chrom_names(vcf.df, ref.genome, name.of.VCF = NULL)
}
\arguments{
\item{vcf.df}{A VCF as a data frame with a \code{CHROM} column.}

\item{ref.genome}{A BSgenome object (e.g. \code{BSgenome.Hsapiens.UCSC.hg38}).}

\item{name.of.VCF}{Name of the VCF file (for warning/error messages).}
}
\value{
A character vector of chromosome names that can be used as a
replacement for \code{vcf.df$CHROM}. Errors if reconciliation is impossible.
}
\description{
Harmonizes the \code{CHROM} column of \code{vcf.df} with the chromosome names used
by \code{ref.genome} (a BSgenome object). Adds or strips the \code{chr} prefix as
needed; for human and mouse, translates \code{23/24} or \code{20/21} to \code{X/Y}.
}
\keyword{internal}
