% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_names.R
\name{check_and_fix_chrom_names_for_trans_ranges}
\alias{check_and_fix_chrom_names_for_trans_ranges}
\title{Check and, if possible, correct the chromosome names in a trans.ranges table}
\usage{
check_and_fix_chrom_names_for_trans_ranges(
  trans.ranges,
  vcf.df,
  ref.genome,
  name.of.VCF = NULL
)
}
\arguments{
\item{trans.ranges}{A \code{data.table} of transcript ranges (see
\link{trans.ranges}).}

\item{vcf.df}{A VCF as a data frame with a \code{CHROM} column.}

\item{ref.genome}{A BSgenome object used to determine organism.}

\item{name.of.VCF}{Name of the VCF file.}
}
\value{
A character vector of chromosome names that can be used as a
replacement for \code{trans.ranges$chrom}. Errors if reconciliation is
impossible.
}
\description{
Harmonizes \code{trans.ranges$chrom} with the chromosome naming used by
\code{vcf.df$CHROM}, adding or stripping \code{chr} as needed and translating
organism-specific numeric X/Y encodings.
}
\keyword{internal}
