% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_strand.R
\name{is_grch37}
\alias{is_grch37}
\title{Is this reference genome GRCh37 (1000 Genomes hs37d5)?}
\usage{
is_grch37(x)
}
\arguments{
\item{x}{A BSgenome object or a character identifier.}
}
\value{
A single logical value, \code{TRUE} if \code{x} identifies that genome,
otherwise \code{FALSE}.
}
\description{
Is this reference genome GRCh37 (1000 Genomes hs37d5)?
}
\keyword{internal}
