% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justify_indel.R
\name{justify_indel}
\alias{justify_indel}
\title{Move the notional position of a deletion or insertion as far left as
possible.}
\usage{
justify_indel(long_str, short_str, pos, expected_delta = NULL)
}
\arguments{
\item{long_str}{A single character string}

\item{short_str}{A single character string.}

\item{pos}{An integer; see the description}

\item{expected_delta}{A single string; see the description}
}
\value{
A list with elements
\itemize{
\item leftmost_pos
\item del_string
\item del_string_plus del_string plus one base to the left the poisiton of that
base to the left is leftmost_pos - 1
\item edge_warning See the third example. We are not sure how far to the left we
could have moved the indel if we had more context on the left.
\item error if non-NULL we did not see the expected value for
expected_delta
}
}
\description{
short_string should be generated by a deletion in long-str at
(1-based) position pos. If supplied, the expected_delta should be the deleted
sequence. If expected_delta is not NULL, the function checks this and aborts
if this is not the case.

Dually, pos is the 1-based position in short string in front o which one can
make an insertion to get long_str.

This funcition moves pos as far to left as possible so that a deletion at
that position still results in an edit of long_str to generate short_str.
}
\examples{
justify_indel("CAAAG", "CAAG", pos = 2, expected_delta = "A")
justify_indel("CAAAG", "CAAG", pos = 3, expected_delta = "A")
justify_indel("CACAG", "CAG", pos = 3, expected_delta = "CA")

}
