% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcprefix_fast.R
\name{lcprefix_fast}
\alias{lcprefix_fast}
\title{Compute length of longest common prefix of two strings}
\usage{
lcprefix_fast(a, b)
}
\arguments{
\item{a}{A single character string.}

\item{b}{A single character string.}
}
\value{
An integer: the number of leading characters that match.
}
\description{
Fast alternative to \code{Biostrings::lcprefix} that avoids
S4 method dispatch overhead.
}
\keyword{internal}
