% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_helpers.R
\name{revc}
\alias{revc}
\title{Reverse complement a character vector of DNA strings}
\usage{
revc(x)
}
\arguments{
\item{x}{A character vector of DNA sequences.}
}
\value{
A character vector of reverse complements.
}
\description{
Thin vectorized wrapper around
\link[Biostrings:reverse]{Biostrings::reverseComplement()}. Handles IUPAC
ambiguity codes.
}
