% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_helpers.R
\name{revc_sbs96}
\alias{revc_sbs96}
\title{Reverse complement strings that represent stranded SBSs}
\usage{
revc_sbs96(mutstring)
}
\arguments{
\item{mutstring}{A character vector of 4-letter strings.}
}
\value{
A character vector the same length as \code{mutstring} containing
the reverse-complemented strings.
}
\description{
Input is a 4-character string where characters 1-3 are the trinucleotide
context and character 4 is the ALT (e.g. \code{"AATC"} = \code{AAT>ACT}).
Returns the reverse complement of the first 3 characters concatenated
with the reverse complement of the last character,
e.g. \code{"AATC"} returns \code{"ATTG"}.
}
\keyword{internal}
