% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_attr.R
\name{stop_if_counts_or_density_illegal}
\alias{stop_if_counts_or_density_illegal}
\title{Validate a counts_or_density argument}
\usage{
stop_if_counts_or_density_illegal(counts_or_density)
}
\value{
\code{NULL}, invisibly. Called for its side effect of raising an
error when \code{counts_or_density} is not a legal value.
}
\description{
Validate a counts_or_density argument
}
\keyword{internal}
