% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_utils.R
\name{stop_if_region_illegal}
\alias{stop_if_region_illegal}
\title{Validate a \code{region} argument}
\usage{
stop_if_region_illegal(region)
}
\arguments{
\item{region}{Character string to check.}
}
\value{
\code{NULL} invisibly; raises an error if \code{region} is not one of
\code{"genome"}, \code{"exome"}, \code{"transcript"}, \code{"unknown"}.
}
\description{
Validate a \code{region} argument
}
\keyword{internal}
