% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_strand.R
\name{is_grch38}
\alias{is_grch38}
\title{Is this reference genome GRCh38 (UCSC hg38)?}
\usage{
is_grch38(x)
}
\arguments{
\item{x}{A BSgenome object or a character identifier.}
}
\description{
Is this reference genome GRCh38 (UCSC hg38)?
}
\keyword{internal}
