% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_strand.R
\name{is_grcm38}
\alias{is_grcm38}
\title{Is this reference genome GRCm38 (UCSC mm10)?}
\usage{
is_grcm38(x)
}
\arguments{
\item{x}{A BSgenome object or a character identifier.}
}
\description{
Is this reference genome GRCm38 (UCSC mm10)?
}
\keyword{internal}
