% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_helpers.R
\name{revc_dbs144}
\alias{revc_dbs144}
\title{Reverse-complement stranded DBS144 4-character class strings}
\usage{
revc_dbs144(x)
}
\arguments{
\item{x}{A character vector of 4-letter strings.}
}
\description{
Input is a 4-character string where characters 1-2 are the REF dinucleotide
and characters 3-4 are the ALT (e.g. \code{"AATC"} = \code{AA>TC}).
}
\keyword{internal}
