% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_helpers.R
\name{revc_sbs96}
\alias{revc_sbs96}
\title{Reverse-complement stranded SBS96 4-character class strings}
\usage{
revc_sbs96(x)
}
\arguments{
\item{x}{A character vector of 4-letter strings.}
}
\description{
Input is a 4-character string where characters 1-3 are the trinucleotide
context and character 4 is the ALT (e.g. \code{"AATC"} = \code{AAT>ACT}).
}
\keyword{internal}
